+++ tests/test_scan_context.py
@@ -23,8 +23,8 @@ from dvr_scan.motion_detector import MotionDetectorCNT, MotionDetectorCudaMOG2
 TRAFFIC_CAMERA_ROI = [631, 532, 210, 127]
 
 TRAFFIC_CAMERA_EVENTS = [
-    (9, 148),
-    (358, 490),
+    (9, 149),
+    (358, 491),
     (542, 576),
 ]
 
@@ -32,62 +32,51 @@ TRAFFIC_CAMERA_EVENTS = [
 CUDA_EVENT_TOLERANCE = 1
 
 TRAFFIC_CAMERA_EVENTS_TIME_PRE_5 = [
-    (3, 148),
-    (352, 490),
+    (3, 149),
+    (352, 491),
     (536, 576),
 ]
 
 # Last event still ends on end of video even though we specified to include 40 frames extra.
 TRAFFIC_CAMERA_EVENTS_TIME_POST_40 = [
-    (9, 138),
-    (358, 480),
+    (9, 139),
+    (358, 481),
     (542, 576),
 ]
 
 # Even though the first frame contains motion, the first frame we can actually detect it on is
 # the second frame (index 1).
 TRAFFIC_CAMERA_EVENTS_CNT = [
-    (1, 148),
-    (364, 490),
+    (1, 149),
+    (364, 491),
     (543, 576),
 ]
 
 # Small ROI for faster test execution.
 CORRUPT_VIDEO_ROI = [0, 0, 32, 32]
 CORRUPT_VIDEO_EVENTS = [
-    (152, 365),
+    (152, 366),
 ]
 
 
 def test_scan_context(traffic_camera_video):
     """Test functionality of ScanContext with default parameters (DetectorType.MOG2)."""
-
     sctx = ScanContext([traffic_camera_video])
     sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
     sctx.set_event_params(min_event_len=4, time_pre_event=0)
-
     event_list = sctx.scan_motion()
-
-    assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS)
-    # Remove duration, check start/end times.
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     assert event_list == TRAFFIC_CAMERA_EVENTS
-
-    # TODO: Add check for duration (should be end - start + 1).
 
 
 @pytest.mark.skipif(not MotionDetectorCudaMOG2.is_available(), reason="CUDA module not available.")
 def test_scan_context_cuda(traffic_camera_video):
     """ Test functionality of ScanContext with the DetectorType.MOG2_CUDA. """
-
     sctx = ScanContext([traffic_camera_video])
     sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
     sctx.set_event_params(min_event_len=4, time_pre_event=0)
-
     event_list = sctx.scan_motion(detector_type=DetectorType.MOG2_CUDA)
-
     assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS)
-    # Remove duration, check start/end times.
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     for i, event in enumerate(event_list):
         start_matches = abs(event[0] - TRAFFIC_CAMERA_EVENTS[i][0]) <= CUDA_EVENT_TOLERANCE
@@ -100,29 +89,20 @@ def test_scan_context_cuda(traffic_camera_video):
 @pytest.mark.skipif(not MotionDetectorCNT.is_available(), reason="CNT algorithm not available.")
 def test_scan_context_cnt(traffic_camera_video):
     """ Test basic functionality of ScanContext using the CNT algorithm. """
-
     sctx = ScanContext([traffic_camera_video])
     sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
     sctx.set_event_params(min_event_len=3, time_pre_event=0)
-
     event_list = sctx.scan_motion(DetectorType.CNT)
-
-    assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS_CNT)
-    # Remove duration, check start/end times.
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     assert event_list == TRAFFIC_CAMERA_EVENTS_CNT
 
 
 def test_pre_event_shift(traffic_camera_video):
     """ Test setting time_pre_event. """
-
     sctx = ScanContext([traffic_camera_video])
     sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
     sctx.set_event_params(min_event_len=4, time_pre_event=6)
-
     event_list = sctx.scan_motion()
-
-    assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS_TIME_PRE_5)
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     assert event_list == TRAFFIC_CAMERA_EVENTS_TIME_PRE_5
 
@@ -130,14 +110,10 @@ def test_pre_event_shift(traffic_camera_video):
 def test_pre_event_shift_with_frame_skip(traffic_camera_video):
     """ Test setting time_pre_event when using frame_skip. """
     for frame_skip in range(1, 6):
-
         sctx = ScanContext([traffic_camera_video], frame_skip=frame_skip)
         sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
         sctx.set_event_params(min_event_len=4, time_pre_event=6)
-
         event_list = sctx.scan_motion()
-
-        assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS_TIME_PRE_5)
         event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
         # The start times should not differ from the ground truth (non-frame-skipped) by the amount
         # of frames that we are skipping. End times can vary more since the default value of
@@ -157,7 +133,6 @@ def test_post_event_shift(traffic_camera_video):
     sctx.set_event_params(min_event_len=4, time_pre_event=0, time_post_event=40)
 
     event_list = sctx.scan_motion()
-
     assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS_TIME_POST_40)
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     assert all([x == y for x, y in zip(event_list, TRAFFIC_CAMERA_EVENTS_TIME_POST_40)])
@@ -166,14 +141,10 @@ def test_post_event_shift(traffic_camera_video):
 def test_post_event_shift_with_frame_skip(traffic_camera_video):
     """ Test setting time_post_event. """
     for frame_skip in range(1, 6):
-
         sctx = ScanContext([traffic_camera_video], frame_skip=frame_skip)
-
         sctx.set_detection_params(roi=TRAFFIC_CAMERA_ROI)
         sctx.set_event_params(min_event_len=4, time_post_event=40)
-
         event_list = sctx.scan_motion()
-
         assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS_TIME_POST_40)
         event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
         # The calculated end times should not differ by more than frame_skip from the ground truth.
@@ -182,7 +153,6 @@ def test_post_event_shift_with_frame_skip(traffic_camera_video):
             for x, y in zip(event_list, TRAFFIC_CAMERA_EVENTS_TIME_POST_40)
         ]), "Comparison failure when frame_skip = %d" % (
             frame_skip)
-
         # The calculated end times must always be >= the ground truth's frame number, otherwise
         # we may be discarding frames containing motion due to skipping them.
         assert all([x[1] >= y[1] for x, y in zip(event_list, TRAFFIC_CAMERA_EVENTS_TIME_POST_40)
@@ -194,10 +164,7 @@ def test_decode_corrupt_video(corrupt_video):
     """Ensure we can process a video with a single bad frame."""
     sctx = ScanContext([corrupt_video])
     sctx.set_detection_params(roi=CORRUPT_VIDEO_ROI)
-
     event_list = sctx.scan_motion()
-
-    assert len(event_list) == len(CORRUPT_VIDEO_EVENTS)
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
     assert event_list == CORRUPT_VIDEO_EVENTS
 
@@ -209,10 +176,8 @@ def test_start_end_time(traffic_camera_video):
     sctx.set_event_params(min_event_len=4, time_pre_event=0)
     sctx.set_video_time(start_time=200, end_time=500)
     event_list = sctx.scan_motion()
-    # The set duration should only cover the middle event.
-    assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS[1:2])
-    # Remove duration, check start/end times.
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
+    # The set duration should only cover the middle event.
     assert event_list == TRAFFIC_CAMERA_EVENTS[1:2]
 
 
@@ -223,8 +188,6 @@ def test_start_duration(traffic_camera_video):
     sctx.set_event_params(min_event_len=4, time_pre_event=0)
     sctx.set_video_time(start_time=200, duration=300)
     event_list = sctx.scan_motion()
-    # The set duration should only cover the middle event.
-    assert len(event_list) == len(TRAFFIC_CAMERA_EVENTS[1:2])
-    # Remove duration, check start/end times.
     event_list = [(event[0].frame_num, event[1].frame_num) for event in event_list]
+    # The set duration should only cover the middle event.
     assert event_list == TRAFFIC_CAMERA_EVENTS[1:2]